Add system test for spectral TE estimation
Add system test on matlab test data. Save results and debug
information in the dev folder. Add script for plotting of
results.

The system test loops over all scales for the test data set.
This should only yield significant results for the 7th scale,
for which information transfer was simulated. This is not
yet working in this commit.