Move all analysis parameters to options dictionary
Move all analysis parameters (lags, taus, stats options, forced
conditional variables, etc.) to a single dictionary. The dictionary
is passed to the class methods performing the analysis instead of
passing the parameters upon instantiation of algorithm classes.

This moves all parameters into one place for more consistency and
better usability. The analysis_options dict is then added to the
algorithm class and returned to the user. All parameters are set to
their default values throughought the analysis to document the
parameters used for analysis.